requirements.txt
pypdf2
sentence-transformers
faiss-cpu
fastapi
uvicorn
